examples/doppler_shift/frequency_changer/pyproject.toml
[tool.poetry]
name = "python-reservator"
version = "0.1.0"
description = ""
authors = ["William Dion <willcar.d@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.0"
lxml = "^4.6.2"
bottle = "^0.12.19"
bottle-cors = "^0.1.5"
python-dotenv = "^0.19.1"
skyfield = "^1.39"



[tool.poetry.dev-dependencies]
pylint = "^2.6.0"

[tool.pylint.messages_control]
disable = [
    "missing-function-docstring",
    "missing-module-docstring",
    "missing-class-docstring",
    "no-self-use",
    "too-many-arguments",
    "import-error",
    "too-few-public-methods"
]

max-line-length=150

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
